Rename/repackage bisq.cli.{app.Bisq=>CliMain}
